pylampo: add deamon functionality in the library
