+++ tools/pylampo/pylampo_client/__init__.py
@@ -1,3 +1,7 @@
 __version__ = "0.1.0"
+import os
+
+if os.getenv('USE_LNPROTOTEST_DEPENDENCIES', 'false').lower() == 'true':
+    from .deamon import start_bitcoind, lampocli_check, start_lampo, reserve_port, wait_for
 
 from .client import LampoClient
